+++ tests/conftest.py
@@ -31,8 +31,6 @@ class DictDiffer:
 
 
 def pytest_assertrepr_compare(op: str, left: MkosiConfig, right: Dict[str, Any]) -> List[str]:
-    if not isinstance(right, dict):
-        return [] # type: ignore
     for r in right.values():
         if not isinstance(vars(r), dict):
             return ["Invalid datatype"]