+++ NEWS.md
@@ -11,6 +11,10 @@
   on this config file to configure other interfaces, you'll have to re-create it
   with the correct match and a lower initial number in the filename to make sure networkd will keep configuring your
   interface, e.g. via the `mkosi.skeleton` or `mkosi.extra` trees or a `mkosi.postinst` script.
+- The kernel-install script for building unified kernel images was removed. From v13 onwards, on systems using kernel-install,
+  mkosi won't automatically build new unified kernel images anymore when a kernel is updated or installed. To keep this
+  behavior, you can install the kernel-install script manually via a skeleton tree. The original script can be found
+  [here](https://github.com/systemd/mkosi/blob/3798eb0c2ebcdf7dac207a559a3cb5a65cdb77b0/mkosi/resources/dracut_unified_kernel_install.sh).
 
 ## v12